Avoid creating results with numpy scalars (re: NEP 51)

+++ networkx/algorithms/centrality/current_flow_betweenness_subset.py
@@ -116,7 +116,7 @@ def current_flow_betweenness_centrality_subset(
         nb = 2.0
     for v in H:
         betweenness[v] = betweenness[v] / nb + 1.0 / (2 - n)
-    return {ordering[k]: v for k, v in betweenness.items()}
+    return {ordering[k]: float(v) for k, v in betweenness.items()}
 
 
 @not_implemented_for("directed")
@@ -221,6 +221,6 @@ def edge_current_flow_betweenness_centrality_subset(
             i = mapping[ss]
             for tt in targets:
                 j = mapping[tt]
-                betweenness[e] += 0.5 * np.abs(row[i] - row[j])
+                betweenness[e] += float(0.5 * np.abs(row[i] - row[j]))
         betweenness[e] /= nb
     return {(ordering[s], ordering[t]): v for (s, t), v in betweenness.items()}
